pyproject.toml
[project]
name = "fh-franken-ui"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Daniel Roy Greenfeld", email = "62857+pydanny@users.noreply.github.com" },
    { name = "Audrey Roy Greenfeld", email = "62857+audreyfeldroy@users.noreply.github.com" }

]
dependencies = [
    "python-fasthtml@git+https://github.com/AnswerDotAI/fasthtml"
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.5.1",
    "ipykernel>=6.29.5",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/fh_franken_ui"]
